fix: Change linesCleared type from Int to Long

+++ core/data/src/commonMain/kotlin/com/yet/tetris/data/model/GameRecordDto.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class GameRecordDto(
     val id: String,
     val score: Long,
-    val linesCleared: Int,
+    val linesCleared: Long,
     val difficulty: DifficultyDto,
     val timestamp: Long
 )
