+++ core/data/src/commonMain/kotlin/com/yet/tetris/data/model/GameStateDto.kt
@@ -9,7 +9,7 @@ data class GameStateDto(
     val currentPosition: PositionDto,
     val nextPiece: TetrominoDto,
     val score: Long = 0,
-    val linesCleared: Int = 0,
+    val linesCleared: Long = 0,
     val isGameOver: Boolean = false,
     val isPaused: Boolean = false
 )
